@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 1.0.0+100
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -14,8 +14,8 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   flutter_share_me:
     path: ../
-  image_picker: ^0.6.7+22
-  path_provider: ^1.6.28
+  image_picker: ^0.8.5+3
+  path_provider: ^2.0.10
 
 dev_dependencies:
   flutter_test:
