@@ -5,7 +5,7 @@ homepage: https://github.com/OpenFlutter/flutter_share_me
 publish_to: http://pub.dev
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
