@@ -1,6 +1,5 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_share_me/flutter_share_me.dart';
 
 void main() {
   const MethodChannel channel = MethodChannel('flutter_share_me');
@@ -18,6 +17,6 @@ void main() {
   });
 
   test('getPlatformVersion', () async {
-   // expect(await FlutterShareMe.platformVersion, '42');
+    // expect(await FlutterShareMe.platformVersion, '42');
   });
 }
